% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extend_ncaa_pbp.R
\name{addTeamInfo}
\alias{addTeamInfo}
\title{Add Team Info}
\usage{
addTeamInfo(pbp_df)
}
\arguments{
\item{pbp_df}{the raw play-by-play dataframe, ie the output of getGamePbp
(or scrapeSchoolSeason, scrapeConferenceSeason)}
}
\value{
returns a 17-column data frame, adding columns for year, away (visiting team),
  home (home team), away_conference (ie Pac-12), away_division (ie 1 for Division 1),
  home_conference, and home_division
}
\description{
Adds team information to scraped play-by-play dataframe
}
\examples{
game_pbp <- getGamePbp("https://stats.ncaa.org/game/index/4502166?org_id=674")
addTeamInfo(game_pbp)
\dontrun{
  season_pbp <- scrapeSchoolSeason("Stanford", 2018)
  addteamInfo(season_pbp)
}

}
