% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_ncaa_pbp.R
\name{getGamePbp}
\alias{getGamePbp}
\title{Scrape Game Play-by-Play}
\usage{
getGamePbp(game_info_url)
}
\arguments{
\item{game_info_url}{the stats.ncaa.org page with the game play-by-play,
accessible via baseballr::get_ncaa_schedule_info function}
}
\value{
returns a 10-column data frame with columns for date, location, attendance,
  inning, inning_top_bot, score, batting, fielding, description, game_info_url
}
\description{
Given a game_info_url, returns the play-by-play for that game
}
\examples{
getGamePbp("https://stats.ncaa.org/game/index/4502166?org_id=674")

}
