% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_ncaa_pbp.R
\name{getSchoolMap}
\alias{getSchoolMap}
\title{Get School Map}
\usage{
getSchoolMap()
}
\value{
returns a 6-column data frame with columns for school, conference, school_id,
  year, division, and conference_id
}
\description{
Obtain table mapping school names to conferences,
  school_id, year, division, conference_id
}
\examples{
getSchoolMap()

}
