% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_ncaa_pbp.R
\name{getSchoolSeasonURLs}
\alias{getSchoolSeasonURLs}
\title{Get School Season URLs}
\usage{
getSchoolSeasonURLs(school_name, season_year)
}
\arguments{
\item{school_name}{the name of the school
(must have an exact match in school column of getSchoolMap())}

\item{season_year}{the year of the season to acquire data for
(must have an exact match in the year column of getSchoolMap())}
}
\value{
returns a character vector of URLs for each game
}
\description{
Obtain game_info_url values for all games in a school's season
}
\examples{
getSchoolSeasonURLs("Stanford", 2018)

}
