% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_ncaa_pbp.R
\name{scrapeConferenceSeason}
\alias{scrapeConferenceSeason}
\title{Scrape Conference Season}
\usage{
scrapeConferenceSeason(conference_name, season_year)
}
\arguments{
\item{conference_name}{the name of the conference
(must have an exact match in conference column of getSchoolMap())}

\item{season_year}{the year of the season to acquire data for
(must have an exact match in the year column of getSchoolMap())}
}
\value{
returns a 10-column data frame with columns for date, location, attendance,
  inning, inning_top_bot, score, batting, fielding, description, and game_info_url
}
\description{
Obtain play-by-play data for all games in a season played by all teams in a given conference
}
\examples{
\dontrun{scrapeConferenceSeason("Pac-12", 2018)}

}
