% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_ncaa_pbp.R
\name{scrapeSchoolSeason}
\alias{scrapeSchoolSeason}
\title{Scrape School Season}
\usage{
scrapeSchoolSeason(school_name, season_year)
}
\arguments{
\item{school_name}{the name of the school
(must have an exact match in school column of getSchoolMap())}

\item{season_year}{the year of the season to acquire data for
(must have an exact match in the year column of getSchoolMap())}
}
\value{
returns a 10-column data frame with columns for date, location, attendance,
  inning, inning_top_bot, score, batting, fielding, description, and game_info_url
}
\description{
Obtain play-by-play data for all games in a school's season
}
\examples{
\dontrun{scrapeSchoolSeason("Stanford", 2018)}

}
